Move simple types to a `types` module
